# openapi.yaml
openapi: 3.0.3
info:
  title: SarDag Trading Signals API
  description: |
    Production-ready trading signals API powered by 18 strategies, AI analysis, and real-time data.

    ## Features
    - 🎯 18 Advanced Trading Strategies
    - 🤖 AI-Powered Signal Analysis (3-layer system)
    - 📊 Real-time Market Data (Binance WebSocket)
    - 🔔 Push Notification System
    - 🌍 Traditional Markets Support (Stocks, Gold, Oil)
    - 🧪 Conservative & Aggressive Signal Modes
    - 📈 Historical Backtesting Data
    - 🔐 Circuit Breaker Protection
    - 📝 White-hat Compliance Logging

    ## Base URL
    - Development: `http://localhost:3001`
    - Production: `https://your-domain.vercel.app`

    ## Rate Limiting
    - Standard: 100 requests/minute
    - Authenticated: 500 requests/minute

  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server (port 3001)
  - url: https://your-domain.vercel.app
    description: Production server
...
paths:
...
  /api/conservative-signals:
    post:
      tags:
        - Signals
      summary: Conservative buy signals only
      description: |
        Ultra-safe signals with 9-condition confirmation system.
        Only returns BUY signals with 80%+ confidence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                symbol:
                  type: string
                  example: "ETHUSDT"
              required:
                - symbol
      responses:
        '200':
          description: Conservative signal analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signal'
...
  /api/push/send:
    post:
      tags:
        - Push Notifications
      summary: Send push notification
      description: Send notification to registered devices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                message:
                  type: string
                data:
                  type: object
                tokens:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Notification sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sent:
                    type: integer
...
components:
  schemas:
    Signal:
      type: object
      properties:
        name:
          type: string
          description: Strategy name
          example: "Conservative Buy Signal"
        signal:
          type: string
          enum: [BUY, SELL, WAIT, NEUTRAL]
          description: Trading signal type
        confidence:
          type: number
          minimum: 0
          maximum: 100
          description: Signal confidence percentage
          example: 85
        reason:
          type: string
          description: Human-readable reason for the signal
          example: "9/9 conditions met: Volume spike, RSI oversold, support bounce"
        targets:
          type: array
          items:
            type: number
          description: Take-profit targets
          example: [50000, 52000, 55000]
        stopLoss:
          type: number
          description: Stop-loss price
          example: 48000
        timeframe:
          type: string
          description: Recommended timeframe
          example: "4H"
        indicators:
          type: object
          description: Technical indicator values
          additionalProperties: true
      required:
        - name
        - signal
        - confidence
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for authenticated requests
security:
  - ApiKeyAuth: []